requirements.txt
# Core
Django>=5.0
psycopg[binary]
djangorestframework
djangorestframework-simplejwt

# API Docs
drf-spectacular

# Testing
pytest
pytest-django
pytest-cov
factory-boy